1. 경고 메시지 수정   1)  재 선언된 MAX_PATH 경고 수정: C++이 제공해주는 라이브러리에 있으므로 내가 만든 건 삭제함.   2) size_t -> DWORD 변환 경고 수정: size_t 변수를 int로 형변환함.

+++ Practice/CTempFunction_Client.cpp
@@ -73,7 +73,7 @@ void CTempFunction_Client::init_func(void)
       while (++i <= r_map_response.size())
       {
         std::string msg = "  ";
-        ::WriteFile(pipe_handle, msg.c_str(), strlen(msg.c_str()), &dword_nwrite, NULL);
+        ::WriteFile(pipe_handle, msg.c_str(), (int)strlen(msg.c_str()), &dword_nwrite, NULL);
         std::cout << "작성 == " << msg << std::endl;
         memset(buf, 0x00, BUF_SIZE);
         ::FlushFileBuffers(pipe_handle);
